move type conversion logic to custom types to avoid import error
