+++ lmcache/v1/multiprocess/server.py
@@ -26,7 +26,12 @@ from lmcache.logging import init_logger
 from lmcache.utils import _lmcache_nvtx_annotate
 from lmcache.v1.gpu_connector import lmcache_memcpy_async_d2h, lmcache_memcpy_async_h2d
 from lmcache.v1.memory_management import MemoryFormat, MemoryObj
-from lmcache.v1.multiprocess.custom_types import IPCCacheEngineKey, KVCache, StorageKey
+from lmcache.v1.multiprocess.custom_types import (
+    IPCCacheEngineKey,
+    KVCache,
+    StorageKey,
+    ipc_keys_to_storage_keys,
+)
 from lmcache.v1.multiprocess.mp_storage_manager import MPStorageManager
 from lmcache.v1.multiprocess.mq import MessageQueueServer
 from lmcache.v1.multiprocess.protocol import (
@@ -51,64 +56,6 @@ def list_to_gpu_tensor(lis: list[int], device: torch.device) -> torch.Tensor:
     return torch.frombuffer(array.array("l", lis), dtype=torch.long).to(
         device, non_blocking=True
     )
-
-
-def ipc_keys_to_storage_keys(ipc_keys: list[IPCCacheEngineKey]) -> list[StorageKey]:
-    """
-    Converts a list of IPCCacheEngineKeys to a list of StorageKeys.
-
-    When scheduler calls `lookup`, the corresponding IPCCacheEngineKey will have
-    worker_id = None. In this case, this means "lookup the given model name and chunk
-    hash for ALL workers".
-
-    When worker calls `store` or `retrieve`, the corresponding IPCCacheEngineKey will
-    have worker_id != None. In this case, this means "store/retrieve the given model
-    name and chunk hash for the given worker".
-
-    Args:
-        ipc_keys: List of IPC cache engine keys to convert
-
-    Returns:
-        List of storage keys. If any IPC key has worker_id=None, it will be expanded
-        to one storage key per worker (based on world_size).
-
-    Raises:
-        ValueError: If IPC keys have inconsistent world_size values
-    """
-    if not ipc_keys:
-        return []
-
-    # Validate that all keys have the same world_size
-    world_size = ipc_keys[0].world_size
-    if not all(ipc_key.world_size == world_size for ipc_key in ipc_keys):
-        raise ValueError(
-            "All IPC keys must have the same world_size. Found world_size values:"
-            f" {set(ipc_key.world_size for ipc_key in ipc_keys)}"
-        )
-
-    storage_keys = []
-    for ipc_key in ipc_keys:
-        if ipc_key.worker_id is None:
-            for worker_id in range(ipc_key.world_size):
-                storage_keys.append(
-                    StorageKey(
-                        model_name=ipc_key.model_name,
-                        world_size=ipc_key.world_size,
-                        worker_id=worker_id,
-                        chunk_hash=ipc_key.chunk_hash,
-                    )
-                )
-        else:
-            storage_keys.append(
-                StorageKey(
-                    model_name=ipc_key.model_name,
-                    world_size=ipc_key.world_size,
-                    worker_id=ipc_key.worker_id,
-                    chunk_hash=ipc_key.chunk_hash,
-                )
-            )
-
-    return storage_keys
 
 
 class GPUCacheContext:
